38 Glan Road
"'Yes, we could easily give you a shake-down.'

"'That is very awkward. Could I not come at some more convenient hour?'

"'We have judged it best that you should come late. It is to recompense you for any inconvenience
that we are paying to you, a young and unknown man, a fee which would buy an opinion from the very
heads of your profession. Still, of course, if you would like to draw out of the business, there is
plenty of time to do so.'