102 Tamworth Drive
"It is a noiseless lock," said he. "It is no wonder that it did not wake you. This case, I presume,
contains the coronet. We must have a look at it." He opened the case, and taking out the diadem he
laid it upon the table. It was a magnificent specimen of the jeweller's art, and the thirty-six
stones were the finest that I have ever seen. At one side of the coronet was a cracked edge, where a
corner holding three gems had been torn away.

"Now, Mr. Holder," said Holmes, "here is the corner which corresponds to that which has been so
unfortunately lost. Might I beg that you will break it off."

The banker recoiled in horror. "I should not dream of trying," said he.

"Then I will." Holmes suddenly bent his strength upon it, but without result. "I feel it give a
little," said he; "but, though I am exceptionally strong in the fingers, it would take me all my
time to break it. An ordinary man could not do it. Now, what do you think would happen if I did
break it, Mr. Holder? There would be a noise like a pistol shot. Do you tell me that all this
happened within a few yards of your bed and that you heard nothing of it?"