67 Glan Road
"But it is impossible," said Miss Hunter; "the ladder was not there when the Rucastles went away."

"He has come back and done it. I tell you that he is a clever and dangerous man. I should not be
very much surprised if this were he whose step I hear now upon the stair. I think, Watson, that it
would be as well for you to have your pistol ready."

The words were hardly out of his mouth before a man appeared at the door of the room, a very fat and
burly man, with a heavy stick in his hand. Miss Hunter screamed and shrunk against the wall at the
sight of him, but Sherlock Holmes sprang forward and confronted him.

"You villain!" said he, "where's your daughter?"

The fat man cast his eyes round, and then up at the open skylight.