69 Beaconside
"Pray compose yourself, sir," said Holmes, "and let me have a clear account of who you are and what
it is that has befallen you."

"My name," answered our visitor, "is probably familiar to your ears. I am Alexander Holder, of the
banking firm of Holder & Stevenson, of Threadneedle Street."

The name was indeed well known to us as belonging to the senior partner in the second largest
private banking concern in the City of London. What could have happened, then, to bring one of the
foremost citizens of London to this most pitiable pass? We waited, all curiosity, until with another
effort he braced himself to tell his story.