113 Glan Road
"ARAT," I read.

"And now?" He raised his hand.

"BALLARAT."

"Quite so. That was the word the man uttered, and of which his son only caught the last two
syllables. He was trying to utter the name of his murderer. So and so, of Ballarat."

"It is wonderful!" I exclaimed.