3 Ventnor Terrace
"What shall I do?"

"There is but one thing to do. It must be done at once. You must put this piece of paper which you
have shown us into the brass box which you have described. You must also put in a note to say that
all the other papers were burned by your uncle, and that this is the only one which remains. You
must assert that in such words as will carry conviction with them. Having done this, you must at
once put the box out upon the sundial, as directed. Do you understand?"

"Entirely."

"Do not think of revenge, or anything of the sort, at present. I think that we may gain that by
means of the law; but we have our web to weave, while theirs is already woven. The first
consideration is to remove the pressing danger which threatens you. The second is to clear up the
mystery and to punish the guilty parties."

"I thank you," said the young man, rising and pulling on his overcoat. "You have given me fresh life
and hope. I shall certainly do as you advise."