53 Wilson Gardens
"But how could you guess what the motive was?"

"Had there been women in the house, I should have suspected a mere vulgar intrigue. That, however,
was out of the question. The man's business was a small one, and there was nothing in his house
which could account for such elaborate preparations, and such an expenditure as they were at. It
must, then, be something out of the house. What could it be? I thought of the assistant's fondness
for photography, and his trick of vanishing into the cellar. The cellar! There was the end of this
tangled clue. Then I made inquiries as to this mysterious assistant and found that I had to deal
with one of the coolest and most daring criminals in London. He was doing something in the cellar--
something which took many hours a day for months on end. What could it be, once more? I could think
of nothing save that he was running a tunnel to some other building.