67 Holmefield Avenue
"'And the work?'

"'Is purely nominal.'

"'What do you call purely nominal?'

"'Well, you have to be in the office, or at least in the building, the whole time. If you leave, you
forfeit your whole position forever. The will is very clear upon that point. You don't comply with
the conditions if you budge from the office during that time.'

"'It's only four hours a day, and I should not think of leaving,' said I.