105 Corbet Close
"The first was from Pondicherry, the second from Dundee, and the third from London."

"From East London. What do you deduce from that?"

"They are all seaports. That the writer was on board of a ship."

"Excellent. We have already a clue. There can be no doubt that the probability--the strong
probability--is that the writer was on board of a ship. And now let us consider another point. In
the case of Pondicherry, seven weeks elapsed between the threat and its fulfilment, in Dundee it was
only some three or four days. Does that suggest anything?"