52 Corbet Close
I had forgotten the strange pets which the doctor affected. There was a cheetah, too; perhaps we
might find it upon our shoulders at any moment. I confess that I felt easier in my mind when, after
following Holmes' example and slipping off my shoes, I found myself inside the bedroom. My companion
noiselessly closed the shutters, moved the lamp onto the table, and cast his eyes round the room.
All was as we had seen it in the daytime. Then creeping up to me and making a trumpet of his hand,
he whispered into my ear again so gently that it was all that I could do to distinguish the words:

"The least sound would be fatal to our plans."

I nodded to show that I had heard.

"We must sit without light. He would see it through the ventilator."

I nodded again.

"Do not go asleep; your very life may depend upon it. Have your pistol ready in case we should need
it. I will sit on the side of the bed, and you in that chair."