Crime scene report #165986
"It is, of course, well known to you that in a successful banking business as much depends upon our
being able to find remunerative investments for our funds as upon our increasing our connection and
the number of our depositors. One of our most lucrative means of laying out money is in the shape of
loans, where the security is unimpeachable. We have done a good deal in this direction during the
last few years, and there are many noble families to whom we have advanced large sums upon the
security of their pictures, libraries, or plate.

"Yesterday morning I was seated in my office at the bank when a card was brought in to me by one of
the clerks. I started when I saw the name, for it was that of none other than--well, perhaps even to
you I had better say no more than that it was a name which is a household word all over the earth--
one of the highest, noblest, most exalted names in England. I was overwhelmed by the honour and
attempted, when he entered, to say so, but he plunged at once into business with the air of a man
who wishes to hurry quickly through a disagreeable task.