Crime scene report #713438
Lestrade shrugged his shoulders. "I am afraid that my colleague has been a little quick in forming
his conclusions," he said.

"But he is right. Oh! I know that he is right. James never did it. And about his quarrel with his
father, I am sure that the reason why he would not speak about it to the coroner was because I was
concerned in it."

"In what way?" asked Holmes.

"It is no time for me to hide anything. James and his father had many disagreements about me. Mr.
McCarthy was very anxious that there should be a marriage between us. James and I have always loved
each other as brother and sister; but of course he is young and has seen very little of life yet,
and--and--well, he naturally did not wish to do anything like that yet. So there were quarrels, and
this, I am sure, was one of them."

"And your father?" asked Holmes. "Was he in favour of such a union?"

"No, he was averse to it also. No one but Mr. McCarthy was in favour of it." A quick blush passed
over her fresh young face as Holmes shot one of his keen, questioning glances at her.