Crime scene report #525839
"There's the village," said the driver, pointing to a cluster of roofs some distance to the left;
"but if you want to get to the house, you'll find it shorter to get over this stile, and so by the
foot-path over the fields. There it is, where the lady is walking."

"And the lady, I fancy, is Miss Stoner," observed Holmes, shading his eyes. "Yes, I think we had
better do as you suggest."

We got off, paid our fare, and the trap rattled back on its way to Leatherhead.