Crime scene report #297649
I had been delayed at a case, and it was a little after half-past six when I found myself in Baker
Street once more. As I approached the house I saw a tall man in a Scotch bonnet with a coat which
was buttoned up to his chin waiting outside in the bright semicircle which was thrown from the
fanlight. Just as I arrived the door was opened, and we were shown up together to Holmes' room.

"Mr. Henry Baker, I believe," said he, rising from his armchair and greeting his visitor with the
easy air of geniality which he could so readily assume. "Pray take this chair by the fire, Mr.
Baker. It is a cold night, and I observe that your circulation is more adapted for summer than for
winter. Ah, Watson, you have just come at the right time. Is that your hat, Mr. Baker?"

"Yes, sir, that is undoubtedly my hat."

He was a large man with rounded shoulders, a massive head, and a broad, intelligent face, sloping
down to a pointed beard of grizzled brown. A touch of red in nose and cheeks, with a slight tremor
of his extended hand, recalled Holmes' surmise as to his habits. His rusty black frock-coat was
buttoned right up in front, with the collar turned up, and his lank wrists protruded from his
sleeves without a sign of cuff or shirt. He spoke in a slow staccato fashion, choosing his words
with care, and gave the impression generally of a man of learning and letters who had had ill-usage
at the hands of fortune.

"We have retained these things for some days," said Holmes, "because we expected to see an
advertisement from you giving your address. I am at a loss to know now why you did not advertise."

Our visitor gave a rather shamefaced laugh. "Shillings have not been so plentiful with me as they
once were," he remarked. "I had no doubt that the gang of roughs who assaulted me had carried off
both my hat and the bird. I did not care to spend more money in a hopeless attempt at recovering
them."