Crime scene report #391170
"'Oh, never mind about that, Mr. Wilson!' said Vincent Spaulding. 'I should be able to look after
that for you.'

"'What would be the hours?' I asked.

"'Ten to two.'

"Now a pawnbroker's business is mostly done of an evening, Mr. Holmes, especially Thursday and
Friday evening, which is just before pay-day; so it would suit me very well to earn a little in the
mornings. Besides, I knew that my assistant was a good man, and that he would see to anything that
turned up.