Crime scene report #578042
"That will do," said Holmes. "As to the letters," he continued, glancing over them, "they are very
commonplace. Absolutely no clue in them to Mr. Angel, save that he quotes Balzac once. There is one
remarkable point, however, which will no doubt strike you."

"They are typewritten," I remarked.

"Not only that, but the signature is typewritten. Look at the neat little 'Hosmer Angel' at the
bottom. There is a date, you see, but no superscription except Leadenhall Street, which is rather
vague. The point about the signature is very suggestive--in fact, we may call it conclusive."

"Of what?"

"My dear fellow, is it possible you do not see how strongly it bears upon the case?"