Crime scene report #361911
"Well, I have no connection with any other people who have been making inquiries," said Holmes
carelessly. "If you won't tell us the bet is off, that is all. But I'm always ready to back my
opinion on a matter of fowls, and I have a fiver on it that the bird I ate is country bred."

"Well, then, you've lost your fiver, for it's town bred," snapped the salesman.

"It's nothing of the kind."