Crime scene report #468706
"No, no; the real name," said Holmes sweetly. "It is always awkward doing business with an alias."

A flush sprang to the white cheeks of the stranger. "Well then," said he, "my real name is James
Ryder."

"Precisely so. Head attendant at the Hotel Cosmopolitan. Pray step into the cab, and I shall soon be
able to tell you everything which you would wish to know."

The little man stood glancing from one to the other of us with half-frightened, half-hopeful eyes,
as one who is not sure whether he is on the verge of a windfall or of a catastrophe. Then he stepped
into the cab, and in half an hour we were back in the sitting-room at Baker Street. Nothing had been
said during our drive, but the high, thin breathing of our new companion, and the claspings and
unclaspings of his hands, spoke of the nervous tension within him.