Crime scene report #463722
"Thank you. Pray proceed."

"When my father took over the Horsham property, he, at my request, made a careful examination of the
attic, which had been always locked up. We found the brass box there, although its contents had been
destroyed. On the inside of the cover was a paper label, with the initials of K. K. K. repeated upon
it, and 'Letters, memoranda, receipts, and a register' written beneath. These, we presume, indicated
the nature of the papers which had been destroyed by Colonel Openshaw. For the rest, there was
nothing of much importance in the attic save a great many scattered papers and note-books bearing
upon my uncle's life in America. Some of them were of the war time and showed that he had done his
duty well and had borne the repute of a brave soldier. Others were of a date during the
reconstruction of the Southern states, and were mostly concerned with politics, for he had evidently
taken a strong part in opposing the carpet-bag politicians who had been sent down from the North.

"Well, it was the beginning of '84 when my father came to live at Horsham, and all went as well as
possible with us until the January of '85. On the fourth day after the new year I heard my father
give a sharp cry of surprise as we sat together at the breakfast-table. There he was, sitting with a
newly opened envelope in one hand and five dried orange pips in the outstretched palm of the other
one. He had always laughed at what he called my cock-and-bull story about the colonel, but he looked
very scared and puzzled now that the same thing had come upon himself.

"'Why, what on earth does this mean, John?' he stammered.

"My heart had turned to lead. 'It is K. K. K.,' said I.

"He looked inside the envelope. 'So it is,' he cried. 'Here are the very letters. But what is this
written above them?'