Crime scene report #207810
Holmes had brought up a long thin cane, and this he placed upon the bed beside him. By it he laid
the box of matches and the stump of a candle. Then he turned down the lamp, and we were left in
darkness.

How shall I ever forget that dreadful vigil? I could not hear a sound, not even the drawing of a
breath, and yet I knew that my companion sat open-eyed, within a few feet of me, in the same state
of nervous tension in which I was myself. The shutters cut off the least ray of light, and we waited
in absolute darkness.

From outside came the occasional cry of a night-bird, and once at our very window a long drawn
catlike whine, which told us that the cheetah was indeed at liberty. Far away we could hear the deep
tones of the parish clock, which boomed out every quarter of an hour. How long they seemed, those
quarters! Twelve struck, and one and two and three, and still we sat waiting silently for whatever
might befall.

Suddenly there was the momentary gleam of a light up in the direction of the ventilator, which
vanished immediately, but was succeeded by a strong smell of burning oil and heated metal. Someone
in the next room had lit a dark-lantern. I heard a gentle sound of movement, and then all was silent
once more, though the smell grew stronger. For half an hour I sat with straining ears. Then suddenly
another sound became audible--a very gentle, soothing sound, like that of a small jet of steam
escaping continually from a kettle. The instant that we heard it, Holmes sprang from the bed, struck
a match, and lashed furiously with his cane at the bell-pull.

"You see it, Watson?" he yelled. "You see it?"