Crime scene report #845862
"That is not quite so common, is it? Ah, me! it's a wicked world, and when a clever man turns his
brains to crime it is the worst of all. I think that I have seen enough now, Miss Stoner, and with
your permission we shall walk out upon the lawn."

I had never seen my friend's face so grim or his brow so dark as it was when we turned from the
scene of this investigation. We had walked several times up and down the lawn, neither Miss Stoner
nor myself liking to break in upon his thoughts before he roused himself from his reverie.

"It is very essential, Miss Stoner," said he, "that you should absolutely follow my advice in every
respect."

"I shall most certainly do so."

"The matter is too serious for any hesitation. Your life may depend upon your compliance."

"I assure you that I am in your hands."