Crime scene report #121460
"What a woman--oh, what a woman!" cried the King of Bohemia, when we had all three read this
epistle. "Did I not tell you how quick and resolute she was? Would she not have made an admirable
queen? Is it not a pity that she was not on my level?"

"From what I have seen of the lady she seems indeed to be on a very different level to your
Majesty," said Holmes coldly. "I am sorry that I have not been able to bring your Majesty's business
to a more successful conclusion."

"On the contrary, my dear sir," cried the King; "nothing could be more successful. I know that her
word is inviolate. The photograph is now as safe as if it were in the fire."

"I am glad to hear your Majesty say so."