Crime scene report #375466
Murder took place ten days ago, on July 14th. Coroner report says it was some time between 16:48 and
19:49.

According to the security guard that sat at the entrance to the factory, on the day of the murder no
one came through the front door. So that leaves the back door, which requires an RFID card to open.

With the help of the factory's IT department I managed to cherry-pick the relevant commits from the
access log for all the electronically locked facilities on premise. They're on my personal
'detectives/kpumbinner' branch. But they said the program doesn't have a search button... I'll have
to sift through it to find who used "BACKDOOR_332". I put the log file in the evidence folder to
look through later.