Crime scene report #444147
"'Some little time ago I bought a small place--a very small place--within ten miles of Reading. I
was fortunate enough to discover that there was a deposit of fuller's-earth in one of my fields. On
examining it, however, I found that this deposit was a comparatively small one, and that it formed a
link between two very much larger ones upon the right and left--both of them, however, in the
grounds of my neighbours. These good people were absolutely ignorant that their land contained that
which was quite as valuable as a gold-mine. Naturally, it was to my interest to buy their land
before they discovered its true value, but unfortunately I had no capital by which I could do this.
I took a few of my friends into the secret, however, and they suggested that we should quietly and
secretly work our own little deposit and that in this way we should earn the money which would
enable us to buy the neighbouring fields. This we have now been doing for some time, and in order to
help us in our operations we erected a hydraulic press. This press, as I have already explained, has
got out of order, and we wish your advice upon the subject. We guard our secret very jealously,
however, and if it once became known that we had hydraulic engineers coming to our little house, it
would soon rouse inquiry, and then, if the facts came out, it would be good-bye to any chance of
getting these fields and carrying out our plans. That is why I have made you promise me that you
will not tell a human being that you are going to Eyford to-night. I hope that I make it all plain?'

"'I quite follow you,' said I. 'The only point which I could not quite understand was what use you
could make of a hydraulic press in excavating fuller's-earth, which, as I understand, is dug out
like gravel from a pit.'

"'Ah!' said he carelessly, 'we have our own process. We compress the earth into bricks, so as to
remove them without revealing what they are. But that is a mere detail. I have taken you fully into
my confidence now, Mr. Hatherley, and I have shown you how I trust you.' He rose as he spoke. 'I
shall expect you, then, at Eyford at 11.15.'

"'I shall certainly be there.'

"'And not a word to a soul.' He looked at me with a last long, questioning gaze, and then, pressing
my hand in a cold, dank grasp, he hurried from the room.