Crime scene report #670069
"Just look it up, then."

"There is a train at half-past nine," said I, glancing over my Bradshaw. "It is due at Winchester at
11.30."

"That will do very nicely. Then perhaps I had better postpone my analysis of the acetones, as we may
need to be at our best in the morning."

By eleven o'clock the next day we were well upon our way to the old English capital. Holmes had been
buried in the morning papers all the way down, but after we had passed the Hampshire border he threw
them down and began to admire the scenery. It was an ideal spring day, a light blue sky, flecked
with little fleecy white clouds drifting across from west to east. The sun was shining very
brightly, and yet there was an exhilarating nip in the air, which set an edge to a man's energy. All
over the countryside, away to the rolling hills around Aldershot, the little red and grey roofs of
the farm-steadings peeped out from amid the light green of the new foliage.

"Are they not fresh and beautiful?" I cried with all the enthusiasm of a man fresh from the fogs of
Baker Street.

But Holmes shook his head gravely.