Crime scene report #749943
"It seems to me that you have been very shamefully treated," said Holmes.

"Oh, no, sir! He was too good and kind to leave me so. Why, all the morning he was saying to me
that, whatever happened, I was to be true; and that even if something quite unforeseen occurred to
separate us, I was always to remember that I was pledged to him, and that he would claim his pledge
sooner or later. It seemed strange talk for a wedding-morning, but what has happened since gives a
meaning to it."

"Most certainly it does. Your own opinion is, then, that some unforeseen catastrophe has occurred to
him?"

"Yes, sir. I believe that he foresaw some danger, or else he would not have talked so. And then I
think that what he foresaw happened."

"But you have no notion as to what it could have been?"