Crime scene report #90707
"Certainly, sir," said Baker, who had risen and tucked his newly gained property under his arm.
"There are a few of us who frequent the Alpha Inn, near the Museum--we are to be found in the Museum
itself during the day, you understand. This year our good host, Windigate by name, instituted a
goose club, by which, on consideration of some few pence every week, we were each to receive a bird
at Christmas. My pence were duly paid, and the rest is familiar to you. I am much indebted to you,
sir, for a Scotch bonnet is fitted neither to my years nor my gravity." With a comical pomposity of
manner he bowed solemnly to both of us and strode off upon his way.

"So much for Mr. Henry Baker," said Holmes when he had closed the door behind him. "It is quite
certain that he knows nothing whatever about the matter. Are you hungry, Watson?"

"Not particularly."

"Then I suggest that we turn our dinner into a supper and follow up this clue while it is still
hot."

"By all means."

It was a bitter night, so we drew on our ulsters and wrapped cravats about our throats. Outside, the
stars were shining coldly in a cloudless sky, and the breath of the passers-by blew out into smoke
like so many pistol shots. Our footfalls rang out crisply and loudly as we swung through the
doctors' quarter, Wimpole Street, Harley Street, and so through Wigmore Street into Oxford Street.
In a quarter of an hour we were in Bloomsbury at the Alpha Inn, which is a small public-house at the
corner of one of the streets which runs down into Holborn. Holmes pushed open the door of the
private bar and ordered two glasses of beer from the ruddy-faced, white-aproned landlord.