Crime scene report #831429
"'Sold to Mr. Windigate of the Alpha, at 12s.'"

"What have you to say now?"

Sherlock Holmes looked deeply chagrined. He drew a sovereign from his pocket and threw it down upon
the slab, turning away with the air of a man whose disgust is too deep for words. A few yards off he
stopped under a lamp-post and laughed in the hearty, noiseless fashion which was peculiar to him.

"When you see a man with whiskers of that cut and the 'Pink 'un' protruding out of his pocket, you
can always draw him by a bet," said he. "I daresay that if I had put £100 down in front of him, that
man would not have given me such complete information as was drawn from him by the idea that he was
doing me on a wager. Well, Watson, we are, I fancy, nearing the end of our quest, and the only point
which remains to be determined is whether we should go on to this Mrs. Oakshott to-night, or whether
we should reserve it for to-morrow. It is clear from what that surly fellow said that there are
others besides ourselves who are anxious about the matter, and I should--"

His remarks were suddenly cut short by a loud hubbub which broke out from the stall which we had
just left. Turning round we saw a little rat-faced fellow standing in the centre of the circle of
yellow light which was thrown by the swinging lamp, while Breckinridge, the salesman, framed in the
door of his stall, was shaking his fists fiercely at the cringing figure.

"I've had enough of you and your geese," he shouted. "I wish you were all at the devil together. If
you come pestering me any more with your silly talk I'll set the dog at you. You bring Mrs. Oakshott
here and I'll answer her, but what have you to do with it? Did I buy the geese off you?"