Crime scene report #887464
"When I arrived at the house," continued Holmes, "I at once went very carefully round it to observe
if there were any traces in the snow which might help me. I knew that none had fallen since the
evening before, and also that there had been a strong frost to preserve impressions. I passed along
the tradesmen's path, but found it all trampled down and indistinguishable. Just beyond it, however,
at the far side of the kitchen door, a woman had stood and talked with a man, whose round
impressions on one side showed that he had a wooden leg. I could even tell that they had been
disturbed, for the woman had run back swiftly to the door, as was shown by the deep toe and light
heel marks, while Wooden-leg had waited a little, and then had gone away. I thought at the time that
this might be the maid and her sweetheart, of whom you had already spoken to me, and inquiry showed
it was so. I passed round the garden without seeing anything more than random tracks, which I took
to be the police; but when I got into the stable lane a very long and complex story was written in
the snow in front of me.

"There was a double line of tracks of a booted man, and a second double line which I saw with
delight belonged to a man with naked feet. I was at once convinced from what you had told me that
the latter was your son. The first had walked both ways, but the other had run swiftly, and as his
tread was marked in places over the depression of the boot, it was obvious that he had passed after
the other. I followed them up and found they led to the hall window, where Boots had worn all the
snow away while waiting. Then I walked to the other end, which was a hundred yards or more down the
lane. I saw where Boots had faced round, where the snow was cut up as though there had been a
struggle, and, finally, where a few drops of blood had fallen, to show me that I was not mistaken.
Boots had then run down the lane, and another little smudge of blood showed that it was he who had
been hurt. When he came to the highroad at the other end, I found that the pavement had been
cleared, so there was an end to that clue.

"On entering the house, however, I examined, as you remember, the sill and framework of the hall
window with my lens, and I could at once see that someone had passed out. I could distinguish the
outline of an instep where the wet foot had been placed in coming in. I was then beginning to be
able to form an opinion as to what had occurred. A man had waited outside the window; someone had
brought the gems; the deed had been overseen by your son; he had pursued the thief; had struggled
with him; they had each tugged at the coronet, their united strength causing injuries which neither
alone could have effected. He had returned with the prize, but had left a fragment in the grasp of
his opponent. So far I was clear. The question now was, who was the man and who was it brought him
the coronet?