Crime scene report #839216
"No."

"It was clamped to the floor. Did you ever see a bed fastened like that before?"

"I cannot say that I have."

"The lady could not move her bed. It must always be in the same relative position to the ventilator
and to the rope--or so we may call it, since it was clearly never meant for a bell-pull."

"Holmes," I cried, "I seem to see dimly what you are hinting at. We are only just in time to prevent
some subtle and horrible crime."

"Subtle enough and horrible enough. When a doctor does go wrong he is the first of criminals. He has
nerve and he has knowledge. Palmer and Pritchard were among the heads of their profession. This man
strikes even deeper, but I think, Watson, that we shall be able to strike deeper still. But we shall
have horrors enough before the night is over; for goodness' sake let us have a quiet pipe and turn
our minds for a few hours to something more cheerful."