Crime scene report #893577
"Quite so. That was the word the man uttered, and of which his son only caught the last two
syllables. He was trying to utter the name of his murderer. So and so, of Ballarat."

"It is wonderful!" I exclaimed.

"It is obvious. And now, you see, I had narrowed the field down considerably. The possession of a
grey garment was a third point which, granting the son's statement to be correct, was a certainty.
We have come now out of mere vagueness to the definite conception of an Australian from Ballarat
with a grey cloak."